requirements.txt
asgiref==3.5.2
backports.zoneinfo==0.2.1
click==8.1.3
cmake==3.24.1.1
Django==4.1.3
dlib==19.24.0
face-recognition==1.3.0
face-recognition-models==0.3.0
numpy==1.23.4
opencv-python==4.6.0.66
Pillow==9.2.0
PyMySQL==1.0.2
six==1.16.0
sqlparse==0.4.3
